+++ etils/ecolab/__init__.py
@@ -14,5 +14,5 @@
 
 """Colab public API."""
 
-from etils.ecolab.array_as_img import display_array_as_img
+from etils.ecolab.array_as_img import auto_plot_array
 from etils.ecolab.colab_utils import collapse