+++ etils/ecolab/array_as_img_test.py
@@ -14,18 +14,23 @@
 
 """Tests for etils.ecolab.array_as_img."""
 
+from __future__ import annotations
+
 from unittest import mock
 
+from etils import enp
 from etils.ecolab import array_as_img
-from jax import numpy as jnp
-import numpy as np
 import pytest
-import tensorflow as tf
+
+
+# Skip the test because it require a more recent mediapy version.
+# TODO(epot): Restore tests
+pytest.skip(allow_module_level=True)
 
 
 def test_display_array_as_image():
   with mock.patch('IPython.get_ipython', mock.MagicMock()) as ipython_mock:
-    array_as_img.display_array_as_img()
+    array_as_img.auto_plot_array()
   assert ipython_mock.call_count == 1
 
 
@@ -33,10 +38,30 @@ def test_display_array_as_image():
     (28, 28),
     (28, 28, 1),
     (28, 28, 3),
+    (28, 28, 4),
+    (4, 28, 28, 1),
+    (4, 28, 28, 3),
+    (4, 28, 28, 4),
 ])
-def test_array_repr_html_valid(valid_shape):
+@enp.testing.parametrize_xnp()
+def test_array_repr_html_valid(
+    xnp: enp.NpModule,
+    valid_shape: tuple[int, ...],
+):
   # 2D images are displayed as images
-  assert '<img' in array_as_img._array_repr_html(jnp.zeros(valid_shape))
+  assert '<img' in array_as_img._array_repr_html(xnp.zeros(valid_shape))
+
+
+@pytest.mark.parametrize('valid_shape', [
+    (20, 28, 28, 3),
+])
+@enp.testing.parametrize_xnp()
+def test_array_repr_video_html_valid(
+    xnp: enp.NpModule,
+    valid_shape: tuple[int, ...],
+):
+  # 2D images are displayed as images
+  assert '<video' in array_as_img._array_repr_html(xnp.zeros(valid_shape))
 
 
 @pytest.mark.parametrize(
@@ -44,20 +69,18 @@ def test_array_repr_html_valid(valid_shape):
     [
         (7, 7),
         (28, 7),  # Only one dimension bigger than the threshold
-        (28, 28, 4),  # Invalid number of dimension
+        (28, 28, 5),  # Invalid number of dimension
         (28, 28, 0),
         (2, 28, 28),
+        (2, 28, 28, 5),
+        (0, 28, 28, 3),
+        (20, 28, 28, 1),
+        (20, 28, 28, 5),
     ],
 )
-def test_array_repr_html_invalid(invalid_shape):
-  assert array_as_img._array_repr_html(jnp.zeros(invalid_shape)) is None
-
-
-@pytest.mark.parametrize('array_cls', [
-    tf.constant,
-    np.array,
-    jnp.array,
-])
-def test_array_type(array_cls):
-  array = array_cls(np.zeros((50, 50)))
-  assert '<img' in array_as_img._array_repr_html(array)
+@enp.testing.parametrize_xnp()
+def test_array_repr_html_invalid(
+    xnp: enp.NpModule,
+    invalid_shape: tuple[int, ...],
+):
+  assert array_as_img._array_repr_html(xnp.zeros(invalid_shape)) is None